Retry airgap builds when not found (#20)
As before, kurl will always build new airgap bundles if the database
reports that the installer is new or changed. This adds a check on S3 to
see if the bundle exists and also schedules a build there if not.
Also add ResourceQuota for the kurl namespace so no more than 4 build
jobs will run concurrently.
Also add check for currently running job for the installer and skip if
active.