feat: external add-ons support for past releases
this commit changes the /installer endpoint to also support returning
add-ons and external add-ons for past (older) releases.

a new parameter called "version" was added to the url, if it is not
present the behavior remains the same but when present the endpoint
returns add-ons and external add-ons supported by the provided version.

new url format:

/installer/version/:version

examples:

$ curl https://kurl.sh/installer
	returns the add-ons + external add-ons supported by the latest
	release

$ curl http://kurl.sh/installer/version/v2022.09.28-0
	returns the add-ons + external add-ons supported by the release
	v2022.09.28-0