bug: completely disable new linter
due to a bug in our cloudflare worker all search params are being
trimmed out. there is no way to determine if the user has or not
requested for the new linter to be applied in this scenario.

this commit completely remove the request intercepter and bypasses
everything to the backend directly.